GUI updates.
Export, view, and delete button added for both runs, and tags.
Sreen adjustment is not fully implemented, meaning when the width is too small you can now actually open the side menu.  Apposed to it just disapearing.
we now use [hidden] apposed to *ngIf.  I originally used *ngIf as online [hidden] is treat like the plague online so I shied away from it.  However after further inspection I found out [hidden] works out better for us as their are little realtime updates in the effected apps, this also means we don't have to call for all the data when we get rid of the sidemenu.