Scaling down files, and making more modular.
Removed all deleteRun related code and moved to a seperate file, due to the previous modular update the Run file was starting to get long so I decided Run needs to be seperated.  I plan to move Run related files into a folder, and seperate create, modify and get at a later date.
took all stream lined function for deleteing and wrote them as their own function, and got all queries to refer to the resepected functions as needed.
Also updated to JSON as the method of recieving data.