commented problem area
The html page was giving errors in the console, and since this section will change in the coming week I have commented it out.
Also removed a useless function and call as if became obsolute from the *ngIf to [hidden} update.
finally updated the small name chagne for the date contained in the run class.