delete update tested, and live.
Misread some information so whe I finally was able to test the delete method their was one major issue in front, end and a different on in the back end.
backend, was executing from db, which is allowed in some languages but not python.  fixed by putting it in cursors, and make a function for it as it was getting repetitive in the file.
frontend, wanted subscriptions set up before it would work. Note need to force referesh list after any list adjustments, just taggle ngIf.