Major Edit Update on the frontend
Now use a service as a middle man that determines what is visible, and the state of some text.
Implemented several new subscribe calls, determining the value of text, and most importantly keeping the edit run information uniformed
Minor updates to the Run interface to match up some missmateched database variables.
Added a empty run function, that simply returns the information that is needed for every basic Run object.
Made a second pop-Up menu specifically for edit.
New pop up menu re-enabled, fixed and ready for when the python code is ready.