Edit implemented, plus osme minor fixes:
Added code for editing items within the database.
Added a two show tags (runShow, tagShow), to be implemented in the future for filter, and sort methods.
Updated requirements for tagged, named, and unnamed. now include testing on the borderlines, 0 for tagged, and '' for named.
quick function for commiting the edit to the server.