A bucket load of bug fixes for edit.
Missed yet another missmatch of database to angular variables.
Fixed incorrect path to database.
updateRun no longer needs a run target, as all updates will use the targRun which is held within the runService interface.
Added a setHideEdit function, so when the user clicks to edit another run the window will stay open and just change the data in the fields.
Big edits changes to the edit-menu.html file, so that it now actually sends the information to the respected run variablee.