Branching out file
Rearranged the branching of the files so that the robot controls are seperate from the web service backend code, note they are not delete I have simply moved them and will add the code in along with the recent stream changes.
changed gitignore to account for a venv within Robot_py, along with the files moving
removed index.html as it was never meant to be in the gitHub, and is now obselete for testing as well.