Stream update
after a week and a half of trying different methods finally found a method that gives the results that are acceptable for this project.
__init__ now has flask_socketio, which is used from streaming by making an open connection with the server and client sending images.
Stream contains the code for sending out images at a rate of 1/fps where fps is the number of frames you want per a second.  Currently it is only one camera, however in the future it will need to send out two feeds, for this I hope to implement a class in the future to manage these camera with more ease.