Stream frontend update
app.module adds the new modules for this commit.
app.component adds the code for the stream variable that contains the image as a base64 jpeg
html adds the variable and decoding header needed to show the image that the server recieved.
steram Service recieves the video image, and watches it closely.