Camera Updates
Camera class has been reimplemented with a lot more features to standardise the two cameras that will eventually be added.
temp file has been added to gitignore as the temp file will from now on have files that do not need to be tracked.
tempFileManager has been added to help manage the deletion of the files located inside of the temp folder.
recording and snapping the image of the recording is now implemented.