Starting to work on export
With this I have also started working on tracking tags as this is needed to test export of the shp section.
Also h
as a small stream update that refreshes camera FPS upon updating.
Also a massive overhaul on how the camera is created and deleted. So now it can be turned off reliably, and it cannot try accessing the camera a second time while it already exists.